Don't require genesis accounts data
Since the mainnet's genesis account data is going to be over 100 MiB in
size, it's a good optimization to not require that account data except
for initializing a history node.

History node operators starting their history node for the first time
get a message explaining that they need to obtain a copy of the
mainnet's genesis account and set the `NIMIQ_OVERRIDE_MAINNET_CONFIG`
environment variable to start their history node for the first time.

This cuts down the required genesis config from over 100 MiB to ~4 KiB.

Fixes #3049.