consensus: Fix transaction proofs for non finalized blocks
Fix transaction proofs for non finalized blocks in
`prove_transaction` which is called when a transaction proof is
requested without specifying a block. In this case if the transaction
is found to be within the current batch (not finalized) it was
incorrectly generating the proof against the head of the blockchain
instead of the proving block which in this case is the block that
included the transaction.
This was a condition that was overseen in #1517.