Correct web-client vesting contract data parsing
Now takes the transaction's actual recipient data, instead of the internal representation of a vesting contract's `CreationTransactionData`. The internal representation uses `varints` for the serialization of the contract's `startTime` and `timeStep`, but the transaction recipient data doesn't. Since the transaction recipient data also has formats that doesn't include the contracts `totalAmount`, this field is required now to be passed to the parsing method.